requirements.txt
arrow==0.8.0
asciinema==0.9.8
astroid==1.1.1
awscli==1.3.21
backports.ssl-match-hostname==3.4.0.2
bcdoc==0.12.2
beautifulsoup4==4.4.1
binaryornot==0.4.0
blinker==1.3
botocore==0.55.0
certifi==2015.4.28
cffi==1.6.0
chardet==2.3.0
click==6.6
colorama==0.2.5
conda==4.0.5
conda-env==2.4.2
cookiecutter==1.4.0
coverage==3.7.1
cryptography==1.3.1
cssmin==0.2.0
cv2==1.0
decorator==4.0.10
dj-static==0.0.6
Django==1.6.1
django-mobi==0.1.7
docutils==0.11
enum34==1.1.1
Flask==0.11.1
Flask-Assets==0.10
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-PyMongo==0.4.1
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.4
Flask-WTF==0.9.5
fn==0.4.3
funcsigs==0.4
future==0.15.2
geopy==1.11.0
gevent==1.0.2
glob2==0.4.1
gmpy==1.17
gmpy2==2.0.7
gnureadline==6.3.3
greenlet==0.4.7
grequests==0.2.0
gspread==0.2.3
gunicorn==19.1.1
heroku==0.1.4
httplib2==0.9.2
idna==2.1
imageio==1.5
ipaddress==1.0.16
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.8
jinja2-time==0.2.0
jmespath==0.4.1
jsonschema==2.4.0
lib601===F14-4
llvmlite==0.8.0
logilab-common==0.61.0
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.5.1
mitmproxy==0.11.3
mock==1.0.1
moviepy==0.2.2.11
MySQL-python==1.2.5
Naked==0.1.31
nose==1.3.4
nose2==0.4.7
numba==0.22.1
numpy==1.11.1
oauth2client==1.4.9
opentuner==0.5.0
ordereddict==1.1
paillier==0.1
pbkdf2==1.3
pbr==1.10.0
pdfkit==0.5.0
Pillow==3.1.0
poyo==0.4.0
protobuf==3.0.0b3
ptyprocess==0.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5
pydub==0.16.4
Pygments==2.0.2
pylint==1.2.1
pymongo==3.0.3
pynvvm==1.0.2
pyOpenSSL==16.0.0
pyparsing==2.0.3
PyRSS2Gen==1.0.0
pysqlite==2.8.2
pytesseract==0.1.6
python-dateutil==2.2
python-dotenv==0.3.0
python-ldap==2.4.20
pytz==2015.2
PyYAML==3.11
pyzmq==14.4.1
readchar==1.1.0
requests==2.7.0
robobrowser==0.5.3
rsa==3.1.2
scipy==0.17.1
shellescape==3.4.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
South==1.0.1
SQLAlchemy==1.0.14
static3==0.5.1
stevedore==1.15.0
tensorflow==0.9.0
terminado==0.5
tornado==4.1
tqdm==4.7.6
twilio==5.3.0
urwid==1.3.0
virtualenv==1.11.6
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
webassets==0.10.1
Werkzeug==0.11.10
whichcraft==0.4.0
WTForms==1.0.5
xattr==0.6.4
